HangmangameinC
This is a traditional hangman game in which you have to guess a word letter by letter. For this to work you need to create a text file with name "words.txt" in same directory of this program and add some words in it and each word separated by ' | ' symbol.