Create Equilibrium Index of an Array
Find the sum of the array and then traverse through the array and keep updating the left sum which is initially zero. To get the right sum, subtract the array values from the sum while traversing. Check left sum and right sum at each step. If they are equal, return the current index.