C Program to identify missing Numbers in a given Array
This C Program identifies missing numbers in a given array.
Here is source code of the C Program to identify missing numbers in a given array. The C program is successfully compiled and run on a Linux system. The program output is also shown below.

output should be :

Enter size of array : 6
Enter elements into array : 
1
2
3
5
6
Missing element is : 4