C_Program_to_convert_Number_in_Characters
Number in character conversion: In the c language, using the loop and switch case, we can easily convert numbers to characters. In this program, we take user input and iterate this number until it is 0. We divide it by 10 during the iteration, and the remainder is passed to get the word for the number in the switch case.